+++ src/error.rs
@@ -110,10 +110,10 @@ pub enum Error
 	Internal,
 
 	/// An unknown system error occurred.
-	SystemErr(i32),
+	SystemErr(u32),
 
 	/// An unknown transport error occurred.
-	TransportErr(i32),
+	TransportErr(u32),
 
 	/// Unknown error code
 	///
@@ -122,7 +122,7 @@ pub enum Error
 	/// hidden from the docs because we do not really want to support this and
 	/// to keep prevent additional error types from becoming breaking changes.
 	#[doc(hidden)]
-	Unknown(i32),
+	Unknown(u32),
 }
 impl Error
 {
@@ -131,7 +131,7 @@ impl Error
 	/// This is not an implementation of `From<i32>` because that would make
 	/// the conversion a public part of this crate.
 	#[rustfmt::skip]
-	pub(crate) fn from_code(code: i32) -> Error
+	pub(crate) fn from_code(code: u32) -> Error
 	{
 		match code {
 			0            => panic!("OK result passed as an error"),
@@ -186,7 +186,7 @@ impl From<Error> for io::Error
 	fn from(e: Error) -> io::Error
 	{
 		if let Error::SystemErr(c) = e {
-			io::Error::from_raw_os_error(c)
+			io::Error::from_raw_os_error(c as i32)
 		}
 		else {
 			#[rustfmt::skip]
@@ -258,7 +258,7 @@ impl fmt::Display for Error
 			Error::Ambiguous         => write!(f, "Ambiguous option"),
 			Error::BadType           => write!(f, "Incorrect type"),
 			Error::Internal          => write!(f, "Internal error detected"),
-			Error::SystemErr(c)      => write!(f, "{}", io::Error::from_raw_os_error(c)),
+			Error::SystemErr(c)      => write!(f, "{}", io::Error::from_raw_os_error(c as i32)),
 			Error::TransportErr(c)   => write!(f, "Transport error #{}", c),
 			Error::Unknown(c)        => write!(f, "Unknown error code #{}", c),
 		}
